All the trivial changes for 64-bit support (including commented-out char_to_button_id) Some rakefile tweaking for (hopefully) automated rake push_to_github again

+++ trunk/GosuImpl/Audio/AudioToolboxFile.hpp
@@ -201,7 +201,7 @@ namespace Gosu
             CHECK_OS(ExtAudioFileSeek(file_, 0 + seekOffset_));
         }
         
-        std::size_t readData(void* dest, UInt32 length)
+        std::size_t readData(void* dest, size_t length)
         {
             AudioBufferList abl;
             abl.mNumberBuffers = 1;
